+++ tests/beem/nodes.py
@@ -13,3 +13,7 @@ def get_hive_nodes():
 
 def get_steem_nodes():
     return "https://api.steemit.com"
+
+
+def get_blurt_nodes():
+    return "https://rpc.blurt.world"